requirements.txt
streamlit>=1.30.0
google-generativeai>=0.8.0
pdfplumber>=0.10.0
pandas>=2.0.0
plotly>=5.18.0
scikit-learn>=1.3.0
gTTS>=2.5.0
fpdf2>=2.7.0
